@@ -27,5 +27,5 @@ dev_dependencies:
   pedantic: ^1.10.0
 
 environment:
-  flutter: ">=1.12.13+hotfix.5"
-  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.5.0"
+  sdk: ">=2.14.0 <4.0.0"
