@@ -1,12 +1,12 @@
 name: flutter_fast_lib
-description: A library for quickly creating flutter applications
-version: 0.0.5
+description: A Flutter Fast Lib
+version: 0.1.0
 homepage: https://github.com/AriesHoo/flutter_fast_lib
 email: AriesHoo@126.com
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
@@ -14,47 +14,56 @@ dependencies:
   #  国际化支持
   flutter_localizations:
     sdk: flutter
-  #  状态管理State
-  provider: '>=5.0.0'
-  #  日志打印
+  #  状态管理State dart 2.12
+  provider: '>=6.0.0'
+  #  日志打印 dart 2.12
   logger: '>=1.0.0 <2.0.0'
-  #  网络请求相关dio
-  dio: '>5.0.0'
-  #  下拉刷新
+  #  网络请求相关dio dart 2.15
+  dio: '>=5.2.0'
+  #  下拉刷新 dart 2.12
   pull_to_refresh: '>=2.0.0 <3.0.0'
-  synchronized: '>=3.0.0 <4.0.0'
-  #  本地缓存sp
-  shared_preferences: '>=2.0.5 <4.0.0'
-  #  toast工具类
-  bot_toast: '>=4.0.1'
-  #  统一工具类
-  common_utils: '>=2.0.2'
-  decimal: '<2.0.0'
-  #  自适应/响应式--判定屏幕尺寸
-  adaptive_breakpoints: '>=0.0.5'
-  #  组件生命周期
-  lifecycle: '>=0.4.1 <0.5.0'
-  # WebView
-  webview_flutter: '>=2.3.1'
-  #  设备信息
-  device_info_plus: '>=3.1.1'
-  #  应用包信息
-  package_info_plus: '>=1.3.0'
-  # 文件路径获取
-  path_provider: '>=2.0.2'
+  # dart 2.18
+  synchronized: '>=3.0.1 <4.0.0'
+  #  本地缓存sp dark 2.17
+  shared_preferences: '>=2.0.20'
+  #  toast工具类 dart 2.12
+  bot_toast: '>=4.0.4'
+  #  统一工具类 dart 2.12
+  common_utils: '>=2.1.0'
+  #  decimal: '>=2.0.0'
+  #  自适应/响应式--判定屏幕尺寸 dart 2.17
+  adaptive_breakpoints: '>=0.1.3'
+  #  组件生命周期 dark 2.12
+  lifecycle: '>=0.5.0'
+  # WebView dark 2.18
+  webview_flutter: '>=4.0.0'
+  #  设备信息 dark 2.18
+  device_info_plus: '>=9.0.0'
+  #  应用包信息 dark 2.18
+  package_info_plus: '>=4.0.0'
+  # 文件路径获取 dark 2.17
+  path_provider: '>=2.0.14'
+
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^1.0.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
-# The following section is specific to Flutter.
+# The following section is specific to Flutter packages.
 flutter:
   # This section identifies this Flutter project as a plugin project.
-  # The 'pluginClass' and Android 'package' identifiers should not ordinarily
-  # be modified. They are used by the tooling to maintain consistency when
+  # The 'pluginClass' specifies the class (in Java, Kotlin, Swift, Objective-C, etc.)
+  # which should be registered in the plugin registry. This is required for
+  # using method channels.
+  # The Android 'package' specifies package in which the registered class is.
+  # This is required for using method channels on Android.
+  # The 'ffiPlugin' specifies that native code should be built and bundled.
+  # This is required for using `dart:ffi`.
+  # All these are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
   plugin:
     platforms:
@@ -73,7 +82,7 @@ flutter:
   # https://flutter.dev/assets-and-images/#from-packages
   #
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware
 
   # To add custom fonts to your plugin package, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
