@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_fast_lib/flutter_fast_lib.dart';
+import 'package:flutter_fast_lib_example/main.dart';
 import 'package:flutter_fast_lib_example/view_model/index/index_sample_view_model.dart';
 import 'package:flutter_fast_lib_example/widget/highlight_card_widget.dart';
 
@@ -8,7 +9,7 @@ class TabIndex extends StatefulWidget {
   const TabIndex({Key? key}) : super(key: key);
 
   @override
-  _TabIndexState createState() => _TabIndexState();
+  State<TabIndex> createState() => _TabIndexState();
 }
 
 class _TabIndexState extends State<TabIndex> {
@@ -22,16 +23,8 @@ class _TabIndexState extends State<TabIndex> {
         physics: const ClampingScrollPhysics(),
         slivers: [
           SliverToBoxAdapter(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                AspectRatio(
-                  aspectRatio: 16.0 / 9.0,
-                  child: Container(
-                    color: Theme.of(context).colorScheme.primary,
-                  ),
-                ),
-              ],
+            child: AppBar(
+              title: Text(appString.tabIndex),
             ),
           ),
           SliverList(
