@@ -32,90 +32,90 @@ class FastDialogUtil {
     Color? barrierColor = Colors.black54,
     bool useSafeArea = true,
   }) async {
-    BuildContext _context = context ?? currentContext;
+    BuildContext cxt = context ?? currentContext;
 
     ///默认与当前TargetPlatform保持一致
-    TargetPlatform _platform = Theme.of(_context).platform;
+    TargetPlatform platform = Theme.of(cxt).platform;
 
     ///初步确定是否为Cupertino样式
-    bool _cupertino =
-        _platform == TargetPlatform.iOS || _platform == TargetPlatform.macOS;
+    bool cupertino =
+        platform == TargetPlatform.iOS || platform == TargetPlatform.macOS;
 
     ///根据type类型判断是否忽略平台特性
     if (FastPlatformType.material == platformType) {
-      _cupertino = false;
+      cupertino = false;
     } else if (FastPlatformType.cupertino == platformType) {
-      _cupertino = true;
+      cupertino = true;
     }
 
     ///title
-    Widget? _title = _buildWidget(
-      _context,
+    Widget? titlteWidget = _buildWidget(
+      cxt,
       text: title,
       textColor: titleColor,
       widget: titleWidget,
     );
 
     ///content
-    Widget? _content = _buildWidget(
-      _context,
+    Widget? contentView = _buildWidget(
+      cxt,
       text: content,
       textColor: contentColor,
       widget: contentWidget,
     );
 
-    if (_content != null && !_cupertino) {
-      _content = SingleChildScrollView(
-        child: _content,
+    if (contentView != null && !cupertino) {
+      contentView = SingleChildScrollView(
+        child: contentView,
       );
     }
 
     ///action
-    List<Widget> _listAction = actions ?? [];
+    List<Widget> listAction = actions ?? [];
 
     ///left
-    Widget? _cancel = _buildWidget(
-      _context,
+    Widget? cancelView = _buildWidget(
+      cxt,
       text: cancel,
       textColor: cancelColor,
       widget: cancelWidget,
-      cupertino: _cupertino,
-      onPressed: onCancelPressed ?? () => Navigator.of(_context).pop(0),
+      cupertino: cupertino,
+      onPressed: onCancelPressed ?? () => Navigator.of(cxt).pop(0),
     );
 
     ///right
-    Widget? _ensure = _buildWidget(
-      _context,
+    Widget? ensureView = _buildWidget(
+      cxt,
       text: ensure,
       textColor: ensureColor,
       widget: ensureWidget,
-      cupertino: _cupertino,
-      onPressed: onEnsurePressed ?? () => Navigator.of(_context).pop(1),
+      cupertino: cupertino,
+      onPressed: onEnsurePressed ?? () => Navigator.of(cxt).pop(1),
     );
-    if (_cancel != null) {
-      _listAction.add(_cancel);
+    if (cancelView != null) {
+      listAction.add(cancelView);
     }
-    if (_ensure != null) {
-      _listAction.add(_ensure);
+    if (ensureView != null) {
+      listAction.add(ensureView);
     }
-    if (_cupertino) {
+    if (cupertino) {
       return await showCupertinoDialog<int?>(
-        context: _context,
+        context: cxt,
         barrierLabel: barrierLabel,
         useRootNavigator: useRootNavigator,
         barrierDismissible: barrierDismissible,
         routeSettings: routeSettings,
         builder: (context) {
           return CupertinoAlertDialog(
-            title: _title,
-            content: _content,
-            actions: _listAction,
+            title: titlteWidget,
+            content: contentView,
+            actions: listAction,
           );
         },
       );
     }
     return await showDialog<int?>(
-      context: _context,
+      context: cxt,
       barrierLabel: barrierLabel,
       useRootNavigator: useRootNavigator,
       barrierDismissible: barrierDismissible,
@@ -125,14 +125,14 @@ class FastDialogUtil {
       builder: (context) {
         return AlertDialog(
           clipBehavior: Clip.antiAlias,
-          title: _title,
+          title: titlteWidget,
 
           ///设置Dialog最大宽度
           content: ConstrainedBox(
             constraints: const BoxConstraints(
               maxWidth: 280,
             ),
-            child: _content == null
+            child: contentView == null
                 ? null
                 : Column(
                     mainAxisSize: MainAxisSize.min,
@@ -140,12 +140,12 @@ class FastDialogUtil {
                     children: [
                       ///中间内容部分占据剩余部分保持底部Action Button
                       Flexible(
-                        child: _content,
+                        child: contentView,
                       ),
                     ],
                   ),
           ),
-          actions: _listAction,
+          actions: listAction,
         );
       },
     );
@@ -169,11 +169,11 @@ class FastDialogUtil {
     if (cupertino != null) {
       if (cupertino) {
         return CupertinoButton(
+          onPressed: onPressed,
           child: Text(
             text,
             style: TextStyle(color: textColor),
           ),
-          onPressed: onPressed,
         );
       }
       return TextButton(
