@@ -46,16 +46,16 @@ class FastSpUtil {
 
   /// get object.
   static Map? getObject(String key) {
-    String? _data = _prefs?.getString(key);
-    return (_data == null || _data.isEmpty) ? null : json.decode(_data);
+    String? data = _prefs?.getString(key);
+    return (data == null || data.isEmpty) ? null : json.decode(data);
   }
 
   /// put object list.
   static Future<bool>? putObjectList(String key, List<Object> list) {
-    List<String>? _dataList = list.map((value) {
+    List<String>? dataList = list.map((value) {
       return json.encode(value);
     }).toList();
-    return _prefs?.setStringList(key, _dataList);
+    return _prefs?.setStringList(key, dataList);
   }
 
   /// get obj list.
@@ -72,8 +72,8 @@ class FastSpUtil {
   static List<Map>? getObjectList(String key) {
     List<String>? dataLis = _prefs?.getStringList(key);
     return dataLis?.map((value) {
-      Map _dataMap = json.decode(value);
-      return _dataMap;
+      Map dataMap = json.decode(value);
+      return dataMap;
     }).toList();
   }
 
