@@ -131,78 +131,49 @@ class FastDialog extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final DialogTheme dialogTheme = DialogTheme.of(context);
-    final EdgeInsets effectivePadding = MediaQuery
-        .of(context)
-        .viewInsets +
+    final EdgeInsets effectivePadding = MediaQuery.of(context).viewInsets +
         (insetPadding ??
-            FastManager
-                .getInstance()
-                .dialogMixin
-                .defaultInsetPadding);
-    bool _scrollable =
-        scrollable ?? FastManager
-            .getInstance()
-            .dialogMixin
-            .scrollable;
+            FastManager.getInstance().dialogMixin.defaultInsetPadding);
+    bool canScroll =
+        scrollable ?? FastManager.getInstance().dialogMixin.scrollable;
     double maxBoxWidth =
-        maxWidth ?? FastManager
-            .getInstance()
-            .dialogMixin
-            .maxWidth;
+        maxWidth ?? FastManager.getInstance().dialogMixin.maxWidth;
     double minBoxWidth =
-        minWidth ?? FastManager
-            .getInstance()
-            .dialogMixin
-            .minWidth;
+        minWidth ?? FastManager.getInstance().dialogMixin.minWidth;
 
     double maxBoxHeight =
-        maxHeight ?? FastManager
-            .getInstance()
-            .dialogMixin
-            .maxHeight;
+        maxHeight ?? FastManager.getInstance().dialogMixin.maxHeight;
 
     ///添加内边距
-    Widget _child = Padding(
+    Widget childView = Padding(
       padding: contentPadding ??
-          FastManager
-              .getInstance()
-              .dialogMixin
-              .defaultContentPadding ??
+          FastManager.getInstance().dialogMixin.defaultContentPadding ??
           EdgeInsets.zero,
       child: child,
     );
 
     ///内容可滚动
-    if (_scrollable) {
-      _child = SingleChildScrollView(
+    if (canScroll) {
+      childView = SingleChildScrollView(
         physics: const ClampingScrollPhysics(),
-        child: _child,
+        child: childView,
       );
     }
 
     ///是否显示可关闭按钮
-    bool _closable =
-        closable ?? FastManager
-            .getInstance()
-            .dialogMixin
-            .closable ?? kIsWeb;
+    bool canClose =
+        closable ?? FastManager.getInstance().dialogMixin.closable ?? kIsWeb;
 
     ///添加CloseButton
-    if (_closable) {
-      _child = Stack(
+    if (canClose) {
+      childView = Stack(
         alignment: closeAlignment ??
-            FastManager
-                .getInstance()
-                .dialogMixin
-                .closeAlignment ??
+            FastManager.getInstance().dialogMixin.closeAlignment ??
             AlignmentDirectional.topEnd,
         children: [
-          _child,
+          childView,
           closeButton ??
-              FastManager
-                  .getInstance()
-                  .dialogMixin
-                  .closeButton ??
+              FastManager.getInstance().dialogMixin.closeButton ??
               const CloseButton(),
         ],
       );
@@ -224,31 +195,17 @@ class FastDialog extends StatelessWidget {
           child: Material(
             color: backgroundColor ??
                 dialogTheme.backgroundColor ??
-                Theme
-                    .of(context)
-                    .dialogBackgroundColor,
+                Theme.of(context).dialogBackgroundColor,
             elevation: elevation ??
                 dialogTheme.elevation ??
-                FastManager
-                    .getInstance()
-                    .dialogMixin
-                    .defaultElevation,
+                FastManager.getInstance().dialogMixin.defaultElevation,
             shape: shape ??
                 dialogTheme.shape ??
-                FastManager
-                    .getInstance()
-                    .dialogMixin
-                    .defaultDialogShape,
-            shadowColor: FastManager
-                .getInstance()
-                .dialogMixin
-                .shadowColor,
+                FastManager.getInstance().dialogMixin.defaultDialogShape,
+            shadowColor: FastManager.getInstance().dialogMixin.shadowColor,
             type: MaterialType.card,
             clipBehavior: clipBehavior ??
-                FastManager
-                    .getInstance()
-                    .dialogMixin
-                    .clipBehavior,
+                FastManager.getInstance().dialogMixin.clipBehavior,
             child: ConstrainedBox(
               constraints: BoxConstraints(
                 minWidth: minBoxWidth,
@@ -256,11 +213,10 @@ class FastDialog extends StatelessWidget {
                 maxHeight: maxBoxHeight,
               ),
               child: GestureDetector(
-                onTap: () =>
-                {
+                onTap: () => {
                   ///点击child部分不做关闭操作
                 },
-                child: _child,
+                child: childView,
               ),
             ),
           ),
